polars_compute::arity

Function prim_binary_values

pub fn prim_binary_values<L, R, O, F>(
    lhs: PrimitiveArray<L>,
    rhs: PrimitiveArray<R>,
    op: F,
) -> PrimitiveArray<O>
where
    L: NativeType,
    R: NativeType,
    O: NativeType,
    F: Fn(L, R) -> O,

Apply a binary function to all the values (regardless of nullability) in (lhs, rhs). Combines the validities with a bitand.

May reuse the memory of one of its arguments if possible.