polars_compute

Module gather

Defines take kernel for Array

Modules

binary

binview

bitmap

boolean

fixed_size_list

generic_binary

list

primitive

structure

sublist

Kernels for gathering values contained within lists.

Functions

take_unchecked⚠

Returns a new Array with only indices at indices. Null indices are taken as nulls. The returned array has a length equal to indices.len().