polars_compute

Module unique

Structs

BooleanUniqueKernelState

DictionaryRangedUniqueState

A specialized unique kernel for DictionaryArray for when all values are in a small known range.

PrimitiveRangedUniqueState

A specialized unique kernel for PrimitiveArray for when all values are in a small known range.

Traits

GenericUniqueKernel

A generic unique kernel that selects the generally applicable unique kernel for an Array.

RangedUniqueKernel

Optimized kernel to calculate the unique elements of an array.

SortedUniqueKernel

Kernel to calculate the number of unique elements where the elements are already sorted.