polars_compute::gather

Function take_unchecked

pub unsafe fn take_unchecked(
    values: &dyn Array,
    indices: &IdxArr,
) -> Box<dyn Array>

Returns a new Array with only indices at indices. Null indices are taken as nulls. The returned array has a length equal to indices.len().

Safety

Doesn’t do bound checks